requirements.txt
altair==5.5.0
annotated-types==0.7.0
anyio==4.9.0
asttokens==2.4.1
attrs==23.2.0
beautifulsoup4==4.12.3
blinker==1.9.0
bs4==0.0.2
cachetools==5.5.2
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.8
colorama==0.4.6
cssbeautifier==1.15.4
decorator==5.1.1
distro==1.9.0
djlint==1.36.4
EditorConfig==0.17.0
et-xmlfile==1.1.0
executing==2.0.1
fake-useragent==2.1.0
gitdb==4.0.12
GitPython==3.1.44
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.7
ipython==8.25.0
jedi==0.19.1
Jinja2==3.1.6
jiter==0.9.0
jsbeautifier==1.15.4
json5==0.12.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
langdetect==1.0.9
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
narwhals==1.33.0
numpy==1.26.4
openai==1.70.0
openpyxl==3.1.5
outcome==1.3.0.post0
packaging==24.2
pandas==2.2.2
parso==0.8.4
pathspec==0.12.1
pillow==11.1.0
playwright==1.45.0
plotly==6.0.1
prompt_toolkit==3.0.46
protobuf==5.29.4
pure-eval==0.2.2
pyarrow==19.0.1
pycparser==2.22
pydantic==2.11.2
pydantic_core==2.33.1
pydeck==0.9.1
pyee==11.1.0
Pygments==2.18.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2024.1
PyYAML==6.0.1
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
rpds-py==0.24.0
scipy==1.15.2
selenium==4.21.0
six==1.16.0
smmap==5.0.2
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
stack-data==0.6.3
streamlit==1.44.1
tenacity==9.1.2
toml==0.10.2
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
trio==0.25.1
trio-websocket==0.11.1
typing-inspection==0.4.0
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.1
watchdog==6.0.0
wcwidth==0.2.13
wsproto==1.2.0
